requirements.txt
tweepy
mysql-connector-python
wheel
textblob
pandas
plotly_express
nltk
GetOldTweets3